+++ packages/openclinxr/architecture-rules/src/tsconfig-conventions.test.ts
@@ -1,3 +1,0 @@
-import { describeTsconfigConventionTests } from "./test-suites/tsconfig-conventions.js";
-
-describeTsconfigConventionTests();